Tests/LibWeb/Text/expected/Fetch/fetch-controller-abort.txt

AbortError: Aborted without reason
Abort with a reason
